Move user API in a dedicated file
