+++ src/router.js
@@ -4,10 +4,10 @@ const { parse } = require("querystring");
 const api = require("./api.js");
 const db = require("./db.js");
 const dialogs = require("./dialogs.js");
-const workspaces = require("./workspaces.js");
+const messages = require("./messages.js");
 const scheduler = require("./scheduler.js");
-const NodeCache = require("node-cache");
-const bcrypt = require('bcrypt');
+const users = require("./users.js");
+const workspaces = require("./workspaces.js");
 
 const resourceFolder = {
   ".html": "./public/html",
@@ -22,19 +22,6 @@ const mimeTypes = {
   ".css": "text/css",
   ".ico": "image/x-icon"
 };
-
-var myCache; //server cache
-var ttlCache = 36000; //After 36000 sec (10 hrs), the variables in the cache will be reset
-
-//Functions used to generate the auth token
-//https://stackoverflow.com/questions/8532406/create-a-random-token-in-javascript-based-on-user-details
-var rand = function () {
-  return Math.random().toString(36).substr(2);
-};
-
-var generate_token = function () {
-  return rand() + rand();
-}
 
 var getFilePath = function (request) {
   var extname = String(path.extname(request.url)).toLowerCase();
@@ -72,106 +59,11 @@ var routeStatic = function (request, response) {
 };
 
 var routeApi = function (request, response) {
-  // /api/user
-  if (request.url === "/api/user/login") {
 
-    // GET : retrieve existing user, and send back the username and the token
-    //       used during authentication
-    if (request.method === "POST") {
-      //FIXME : The password can be seen in the request header, should be crypted in the client side
-      var credentials = {
-        username: request.headers.user,
-        password: request.headers.pwd
-      };
-      api.checkCredentialsUser(credentials, function (data) {
-        if (data != false) { //if the user exists
-          bcrypt.compare(credentials['password'], data['password'], function (err, res) {
-            if (res == true) {
-              var generated_token = generate_token();
-              //We get the array tokens in the server cache
-              myCache.get("tokens", function (err, value) {
-                if (!err) {
-                  if (value == undefined) {
-                    //The array in the cache does not exist : Init of the array containing the tokens
-                    myCache.set("tokens", [generated_token], function (err, success) {
-                      if (err) {
-                        response.writeHead(201, { "Content-Type": "application/json" });
-                        response.end();
-                      }
-                    });
-                  } else {
-                    //The array tokens already exists in the cache
-                    var tokens = value;
-                    tokens.push(generated_token);
-                    //Already a token in the server, we add the new generated token in the tokens cache array
-                    myCache.set("tokens", tokens, function (err, success) {
-                      if (err) {
-                        response.writeHead(201, { "Content-Type": "application/json" });
-                        response.end();
-                      }
-                    });
-                  }
-                }
-              });
-              response.writeHead(200, { "Content-Type": "application/json" });
-              response.end(JSON.stringify({ token: generated_token }));
-            } else {
-              response.writeHead(201, { "Content-Type": "application/json" });
-              response.end();
-            }
-          });
-        } else {// The username does not exist in the DB
-          //FIXME : If no user in database, which status code should I return ?
-          response.writeHead(201, { "Content-Type": "application/json" });
-          response.end();
-        }
-      });
-    }
+  if (request.url.startsWith("/api/user")) {
+    users.route(request, response);
   }
-  else if (request.url === "/api/user/logout") {
-    if (request.method === "POST") {
-      myCache.get("tokens", function (err, value) {
-        if (!err) {
-          //We remove this specific token from the server cache
-          var tokens = value;
-          var index_token_to_remove = tokens.indexOf(request.headers.token);
-          tokens.splice(index_token_to_remove, 1);
-          myCache.set("tokens", tokens, function (err, success) {
-            if (err) {
-              response.writeHead(201, { "Content-Type": "application/json" });
-              response.end();
-            }
-          });
-        }
-      });
-      response.writeHead(200, { "Content-Type": "application/json" });
-      response.end();
-    }
-  }
-  else if (request.url === "/api/user") {
-    if (request.method === "POST") {
-      //FIXME : The password can be seen in the request header, should be crypted in the client side
-      var credentials = {
-        username: request.headers.user,
-        password: request.headers.pwd
-      };
-      //See usage of bcrypt library : https://www.npmjs.com/package/bcrypt
-      var saltRounds = 10;
-      bcrypt.hash(credentials['password'], saltRounds, function (err, hash) {
-        credentials['password'] = hash;
-        api.addUser(credentials, function (data) {
-          if (data == false) {//User already existing
-            //FIXME : Status code
-            response.writeHead(201, { "Content-Type": "application/json" });
-            response.end();
-          } else {
-            response.writeHead(200, { "Content-Type": "application/json" });
-            response.end();
-          }
-        });
-      });
-    }
-  }
+
   // /api/config
   else if (request.url === "/api/config") {
 
@@ -241,7 +133,7 @@ var routeApi = function (request, response) {
         if (!infos.ok) {
           response_400(infos, response);
         } else {
-          api.insertObjectInDb("workspaces", infos, function (result) {
+          db.insert("workspaces", infos, function () {
             response.writeHead(302, {
               'Location': `slack://channel?team=${infos.team_id}`
             });
@@ -260,51 +152,11 @@ var routeApi = function (request, response) {
   }
 
   // /api/simple-messages
-  else if (request.url.startsWith("/api/simple-messages")) {
+  else if (request.url.startsWith("/api/messages")) {
+    messages.route(request, response);
+  }
 
-    // GET : retrieve messages
-    if (request.method === "GET") {
-      response.writeHead(200, { "Content-Type": "application/json" });
-      api.listMessages(function (data) {
-        response.write(JSON.stringify(data));
-        response.end();
-      });
-    }
-
-    // DELETE : delete a message
-    else if (request.method === "DELETE") {
-      var regex_delete = /^\/api\/simple-messages\/([^/]+)\/?$/;
-      if (request.url.match(regex_delete) !== null) {
-        var messageId = request.url.match(regex_delete)[1];
-        db.delete("messages", messageId, function (data) {
-          response.write(JSON.stringify(data));
-          response.end();
-        });
-      } else {
-        response.writeHead(404, { "Content-Type": "application/octet-stream" });
-        response.end();
-      }
-    }
-
-    // POST : send a message
-    else if (request.method === "POST") {
-      if (request.url === "/api/simple-messages/send") {
-        let body = "";
-        request.on("data", chunk => {
-          body += chunk.toString();
-        });
-        request.on("end", () => {
-          var parsedBody = parse(body);
-          api.sendSimpleMessage(parsedBody.workspace, parsedBody.channel, parsedBody.message);
-        });
-        response.writeHead(200, { "Content-Type": "application/octet-stream" });
-        response.end();
-      }
-    } else {
-      response.writeHead(404, { "Content-Type": "application/octet-stream" });
-      response.end();
-    }
-  } else {
+  else {
     response.writeHead(404, { "Content-Type": "application/octet-stream" });
     response.end();
   }
@@ -334,7 +186,7 @@ exports.serve = function (request, response) {
   // We check if the token in the request header, is the same that matches the one
   //   which has been saved in the cache server
   if (typeof token !== 'undefined') {
-    myCache.get("tokens", function (err, value) {
+    users.getInCache("tokens", function (err, value) {
       if (!err) {
         if (value == undefined) {
           auth = false;
@@ -377,7 +229,3 @@ exports.serve = function (request, response) {
   }
   return response;
 };
-
-exports.initCache = function () {
-  myCache = new NodeCache({ stdTTL: ttlCache });
-}
